+++ app/helpers/sources_helper.rb
@@ -27,10 +27,16 @@ module SourcesHelper
   end
 
   def source_author_year_tag(source)
-    if source && source.type == 'Source::Bibtex' && source.author_year.present?
-      source.author_year
-    else
-      content_tag(:span, 'Author, year not yet provided for source.', class: [:feedback, 'feedback-thin', 'feedback-warning'])
+    a = content_tag(:span, 'Author, year not yet provided for source.', class: [:feedback, 'feedback-thin', 'feedback-warning'])
+    case source.type
+    when 'Source::Human'
+      source.cached
+    when 'Source::Bibtex'
+      if source.author_year.present?
+        source.author_year
+      end
+    when 'Source::Verbatim'
+      a
     end
   end